Work on moving MIDI editor behavior into `Grid`
 * We were doing all of the handling with the old global `STATE`/`state()` variables even though the `Grid` and `MidiEditor` were set up.  This starts the process of moving all of the state and logic into the grid and view context and out of the global scope.
 * Modify the moved code to adapt it to its new environment
 * Moved many standalone functions in to be methods of `Grid`.  Many of these in turn will be moved to be methods of `MidiEditor` in the future.

+++ engine/src/js.rs
@@ -6,18 +6,18 @@ use wasm_bindgen::prelude::*;
 extern "C" {
     pub fn render_quad(
         canvas_index: usize,
-        x: f32,
-        y: f32,
-        width: f32,
-        height: f32,
+        x: usize,
+        y: usize,
+        width: usize,
+        height: usize,
         class: &str,
     ) -> usize;
     pub fn render_line(
         canvas_index: usize,
-        x1: f32,
-        y1: f32,
-        x2: f32,
-        y2: f32,
+        x1: usize,
+        y1: usize,
+        x2: usize,
+        y2: usize,
         class: &str,
     ) -> usize;
     pub fn get_active_attr(key: &str) -> Option<String>;
